% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{auto_sem}
\alias{auto_sem}
\title{SEM Auto-fitting (Internal)}
\usage{
auto_sem(
  subMEC_lavaan_ready,
  data,
  na.action = na.omit,
  subset = NULL,
  varcov = NULL,
  n = NULL
)
}
\arguments{
\item{subMEC_lavaan_ready}{A list of pre-fitted SEMs in lavaan syntax}

\item{data}{A data frame to fit the given SEM}

\item{na.action}{SEM option}

\item{subset}{SEM option}

\item{varcov}{SEM option}

\item{n}{SEM option}
}
\value{
A lavaan-fitted model list
}
\description{
Takes a lavaan ready list of models and fits all with same basic settings, then produces a fitted list. Not passing sem() arguments yet, and assumes complete data.
}
\keyword{internal}
