% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{calc_A}
\alias{calc_A}
\title{Calculate Matrix A (Internal)}
\usage{
calc_A(SEMfitted)
}
\arguments{
\item{SEMfitted}{A lavaan-fitted SEM}
}
\value{
Matrix A
}
\description{
For a fitted SEM, computes the A information matrix from Vuong (1989).
}
\keyword{internal}
