% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{calc_B}
\alias{calc_B}
\title{Calculate Matrix B (Internal)}
\usage{
calc_B(sc1, sc2, n)
}
\arguments{
\item{sc1}{Mdoel 1 SC Matrix}

\item{sc2}{Model 2 SC Matrix}

\item{n}{Number of observations}
}
\value{
Matrix B
}
\description{
For pairs of SEMS, compute the sample adjusted cross-product B information matrix from Vuong (1989).
}
\keyword{internal}
