% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_vuongmatrices.R
\name{calc_sc}
\alias{calc_sc}
\title{Calculate Score Contribution Matrix (Internal)}
\usage{
calc_sc(SEMfitted)
}
\arguments{
\item{SEMfitted}{A lavaan-fitted SEM}
}
\value{
SC Matrix
}
\description{
For a fitted SEM, computes score contribution matrix (sc).
}
\keyword{internal}
