% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_dagu.R
\name{dagu}
\alias{dagu}
\title{DAG Utility for Model Auto-Generation (Internal)}
\usage{
dagu(LAV_list, path, MEC_only = TRUE)
}
\arguments{
\item{LAV_list}{One or more lavaan syntax models in a list}

\item{path}{A lavaan syntax path}
}
\value{
A model list in dagitty and another in lavaan format
}
\description{
Takes one model and path in lavaan syntax, auto-generates MEC, then gets subMEC_lavaan_ready.
}
\keyword{internal}
