% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mvpt-package}
\alias{mvpt-package}
\title{mvpt: Multiverse Path Test}
\description{
Researchers can often face uncertainty when multiple SEMs equally explain the same data. This can be problematic when an effect (path) between two variables can be modeled in different ways. To help manage estimation uncertainty due to having multiple, competing models, the SEM Multiverse Path Test (MVPT) can auto-generate competing SEMs with the same specified path and test whether this one path changes across models. Significant path value estimation changes will be traceable to alternative model specification changes.
}
\details{
Main functions are \code{\link{mvpt}},
\code{\link{mvptZoom}}, and \code{\link{mvptRank}}.
}
\author{
\strong{Maintainer}: Ronald Yee-Flores \email{rdyeeflores@gmail.edu} (\href{https://orcid.org/0000-0002-7423-1608}{ORCID})

Authors:
\itemize{
  \item Edgar Merkle (\href{https://orcid.org/0000-0001-7158-0653}{ORCID})
  \item Ellen Fitzsimmons (\href{https://orcid.org/0000-0002-0496-067X}{ORCID})
}

}
