% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{mvpt}
\alias{mvpt}
\title{Run MVPT}
\usage{
mvpt(lavaan_input, path, data, showplots = FALSE, MEC_only = TRUE)
}
\arguments{
\item{lavaan_input}{An SEM or list() of SEMs in lavaan syntax.}

\item{path}{The path to be tested. This must also be in lavaan syntax (eg: "Y ~ X").}

\item{data}{A data frame to fit any given SEM, including others that may be auto-generated.}

\item{showplots}{Whether to show a parameter-free figure containing all the compared SEMs (with "M1" first). This figure is for quick inspection of model specification differences.}

\item{MEC_only}{Default is TRUE, but setting to FALSE allows additional models with arrows that feed into the specified path outcome variable.}
}
\value{
An object of class \code{mvpt} containing lavaan-fitted results, figures, and MVPT components.
}
\description{
Using the specification of a single SEM and single path within that SEM, both in \pkg{lavaan} syntax, this function can auto-generate multiple other SEMs with the same single path by only using the graphical features of the given SEM (Verma & Pearl, 1991). Auto-generated models will share the same fit statistics as the given SEM, though suggest differing relationships between variables. This function then uses the supplied data to fit all models using the same settings (maximum likelihood estimation), followed by a chi-square test across models for significant value changes in the specified path. The given model will always be indexed first and appear as "M1" in the output.
}
\details{
If a user already has a set of SEMs to compare, model auto-generation can be bypassed by supplying a list() of SEMs for the lavaan_input argument. Testing would then be limited to these given models.
}
\examples{
\dontrun{

data("UnfairApprais")

lavaan_input <-
  "
  ## latent variables
  Rumi =~ rumi1 + rumi2
  Angr =~ anger1 + anger1
  UnApp =~ unfair1 + unfair2
  ## regressions
  Aggr ~ Angr + Rumi
  Rumi ~ Angr + UnApp
  Angr ~ UnApp
  "

## path test 1
mvpt_path1 <- mvpt(lavaan_input, 
                   path = "Rumi~UnApp", 
                   data = UnfairApprais, 
                   showplots = TRUE)
mvpt_path1 

## path test 2
mvpt_path2 <- mvpt(lavaan_input, 
                   path = "Aggr~Rumi", 
                   data = UnfairApprais, 
                   showplots = TRUE)
mvpt_path2
}
}
