% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvptRank.R
\name{mvptRank}
\alias{mvptRank}
\title{mvpt() Follow-up: Rank Models by Path Value}
\usage{
mvptRank(mvpt_output)
}
\arguments{
\item{mvpt_output}{Output from mvpt().}
}
\description{
Follow-up function to rank and view all models by path value.
}
\examples{
\dontrun{

data("UnfairApprais")

lavaan_input <-
  "
  ## latent variables
  Rumi =~ rumi1 + rumi2
  Angr =~ anger1 + anger1
  UnApp =~ unfair1 + unfair2
  ## regressions
  Aggr ~ Angr + Rumi
  Rumi ~ Angr + UnApp
  Angr ~ UnApp
  "

mvpt_path1 <- mvpt(lavaan_input, 
                   path = "Rumi~UnApp", 
                   data = UnfairApprais, 
                   showplots = TRUE)
mvpt_path1 
mvptRank(mvpt_path1)
}
}
