% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvptZoom.R
\name{mvptZoom}
\alias{mvptZoom}
\title{mvpt() Follow-up: View Single Model}
\usage{
mvptZoom(mvpt_output, index)
}
\arguments{
\item{mvpt_output}{Output from mvpt().}

\item{index}{Model index for list of compared models. Model 1 is always the given model.}
}
\description{
Follow-up function to see lavaan syntax, lavaan summary, and figure for a single SEM.
}
\examples{
\dontrun{

data("UnfairApprais")

lavaan_input <-
  "
  ## latent variables
  Rumi =~ rumi1 + rumi2
  Angr =~ anger1 + anger1
  UnApp =~ unfair1 + unfair2
  ## regressions
  Aggr ~ Angr + Rumi
  Rumi ~ Angr + UnApp
  Angr ~ UnApp
  "

mvpt_path1 <- mvpt(lavaan_input, 
                   path = "Rumi~UnApp", 
                   data = UnfairApprais, 
                   showplots = TRUE)
mvpt_path1 
mvptZoom(mvpt_path1, index = 3)
}
}
