% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mvpt.R
\name{print.mvpt}
\alias{print.mvpt}
\title{Print User-facing Results from mvpt() Object}
\usage{
\method{print}{mvpt}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mvpt}.}

\item{...}{Unused.}
}
\value{
The object \code{x}, invisibly.
}
\description{
Displays a compact summary of an object returned by \code{\link{mvpt}}.
}
