% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-UnfairApprais.R
\docType{data}
\name{UnfairApprais}
\alias{UnfairApprais}
\title{Unfairness Appraisal Dataset}
\format{
A data frame with 7 variables:
\describe{
  \item{unfair1}{Measure of appraised unfairness (item 1).}
  \item{unfair2}{Measure of appraised unfairness (item 2).}
  \item{anger1}{Measure of anger (item 1).}
  \item{anger2}{Measure of anger (item 2).}
  \item{rumi2}{Measure of rumination (item 1).}
  \item{rumi3}{Measure of rumination (item 2).}
  \item{Aggr}{Measure of retaliatory aggression.}
}
}
\source{
Data collected by Ronald D. Yee-Flores.
}
\usage{
UnfairApprais
}
\description{
Data from an unpublished study examining relations among
perceived unfairness, anger, rumination, and retaliatory aggression.
}
\details{
The data is a subset from a larger study, and  intended for 
MVPT demonstration.
}
\examples{
data(UnfairApprais)
str(UnfairApprais)
summary(UnfairApprais)
}
\references{
None. Data have not yet been published.
}
\keyword{datasets}
