% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_vw_core.R
\name{vw_core}
\alias{vw_core}
\title{Calculate Vuong-Wald Core MVP Test Statistic (Internal)}
\usage{
vw_core(SEMfitted_list, path)
}
\arguments{
\item{SEMfitted_list}{A list of lavaan-fitted SEMs}

\item{path}{A lavaan syntax path}
}
\value{
List of lists and more test components
}
\description{
For a list of fitted SEMs and a shared path, a multivariate sampling distribution is used to compare path values between these SEMs.
}
\keyword{internal}
