% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{dagu}
\alias{dagu}
\title{Model Auto-generation DAG Utility (Internal)}
\usage{
dagu(LAV, path)
}
\arguments{
\item{LAV}{A lavaan syntax model}

\item{path}{A lavaan syntax path}
}
\value{
A model list in dagitty and another in lavaan format
}
\description{
Takes one lavaan syntax model and path, auto-generates MEC, then gets subMEC_lavaan_ready.
}
\keyword{internal}
