% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_dagu.R
\name{dagu}
\alias{dagu}
\title{DAG Utility for Model Auto-generation (Internal)}
\usage{
dagu(LAV, path)
}
\arguments{
\item{LAV}{A lavaan syntax model}

\item{path}{A lavaan syntax path}
}
\value{
A model list in dagitty and another in lavaan format
}
\description{
Takes one model and path in lavaan syntax, auto-generates the MEC, then gets subMEC_lavaan_ready.
}
\keyword{internal}
