% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\docType{package}
\name{mvpt-package}
\alias{mvpt-package}
\title{mvpt: Multiverse Path Test}
\description{
Researcher often face uncertainty when multiple SEMs can equally explain the same data. This can be problematic when the effect (path) between two variables can be modeled in different ways. To help manage path value estimation uncertainty due to having competing models, a SEM Multiverse Path (MVP) Test can auto-generate competing SEMs with the same specified path and tests whether this one path is equal across models. Significant path estimation value changes will be traceable to suggested model specification changes.
}
\details{
Main functions are \code{\link{mvpt}},
\code{\link{mvptZoom}}, and \code{\link{mvptRank}}.
}
\author{
\strong{Maintainer}: Ronald Yee-Flores \email{ronaldyeeflores@gmail.edu} (\href{https://orcid.org/0000-0002-7423-1608}{ORCID})

Authors:
\itemize{
  \item Edgar Merkle \email{merklee@missouri.edu} (\href{https://orcid.org/0000-0001-7158-0653}{ORCID})
}

}
