% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\docType{package}
\name{mvpt-package}
\alias{mvpt-package}
\title{mvpt: Multiverse Path Test}
\description{
For individual path estimates within a larger SEM, this is a sensitivity analysis for the robustness of a single path estimate across multiple, automatically-generated, yet equally-fitting SEMs. Significant path estimation value changes will be traceable to suggested model specification changes.
}
\details{
Main functions are \code{\link{mvpt}},
\code{\link{mvptZoom}}, and \code{\link{mvptRank}}.
}
\author{
\strong{Maintainer}: Ronald Yee-Flores \email{ronaldyeeflores@gmail.edu} (\href{https://orcid.org/0000-0002-7423-1608}{ORCID})

Authors:
\itemize{
  \item Edgar Merkle \email{merklee@missouri.edu} (\href{https://orcid.org/0000-0001-7158-0653}{ORCID})
}

}
