% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\docType{package}
\name{mvpt-package}
\alias{mvpt-package}
\title{mvpt: SEM Multiverse Path Test}
\description{
Sensitivity analysis for individual SEM path estimates 
across multiple automatically generated SEMs.
}
\details{
Main functions are \code{\link{mvpt}},
\code{\link{mvptZoom}}, and \code{\link{mvptRank}}.
}
\author{
\strong{Maintainer}: Ronald Yee-Flores \email{ronald.flores@missouri.edu} (\href{https://orcid.org/0000-0002-7423-1608}{ORCID})

Authors:
\itemize{
  \item Edgar Merkle \email{merklee@missouri.edu} (\href{https://orcid.org/0000-0001-7158-0653}{ORCID})
}

}
