% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{mvpt}
\alias{mvpt}
\title{Run MVPT}
\usage{
mvpt(lavaan_input, path, data,
  showplots = FALSE, MEC_only = TRUE)
}
\arguments{
\item{lavaan_input}{A single SEM or list() of SEMs in lavaan syntax. For a single SEM, this function will attempt to auto-generate other models. For a list(), only the listed models will be used for testing.}

\item{path}{The path to be tested. This path must be in lavaan syntax (eg: "Y ~ X") and in any associated model specification.}

\item{data}{A data frame to fit the associated SEMs, including auto-generated SEMs.}

\item{showplots}{Whether to show a parameter-free figure containing all the compared SEMs. User-provided models start at the top-left. This figure is for quick inspection of model specification differences.}

\item{MEC_only}{When given a single SEM, this determines whether model auto-generation goes beyond that model's Markov equivalence class (the space of models with equal fit metrics). In this version of mvpt(), specifying FALSE allows additionally generated models with arrows directly into the outcome variable of the tested path.}
}
\value{
An object of class \code{mvpt} containing lavaan-fitted results, figures, and MVPT components.
}
\description{
Perform an SEM multiverse path test.
}
\details{
\itemize{
  \item This function can auto-generate other SEMs by only using the graphical features of a given SEM (Verma & Pearl, 1991). Auto-generated models will share the same fit statistics, though suggest differing relationships between variables. 
  \item All models are fitted using maximum likelihood estimation, followed by an MVPT (chi-square) across models to determine significant value changes in the tested path. User-provided models will always be indexed first in the output ("M1"). 
  \item If a user already has a set of SEMs to compare, model auto-generation can be bypassed by supplying a list() of SEMs for the lavaan_input argument.
  }
}
\examples{
\dontrun{

data("UnfairApprais")

lavaan_input <-
  "
  ## latent variables
  Rumi =~ rumi1 + rumi2
  Angr =~ anger1 + anger1
  UnApp =~ unfair1 + unfair2
  ## regressions
  Aggr ~ Angr + Rumi
  Rumi ~ Angr + UnApp
  Angr ~ UnApp
  "

## path test 1
mvpt_path1 <- mvpt(lavaan_input, 
                   path = "Rumi~UnApp", 
                   data = UnfairApprais, 
                   showplots = TRUE)
mvpt_path1 

## path test 2
mvpt_path2 <- mvpt(lavaan_input, 
                   path = "Aggr~Rumi", 
                   data = UnfairApprais, 
                   showplots = TRUE)
mvpt_path2
}
}
