% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{mvpt}
\alias{mvpt}
\title{Run MVP Test}
\usage{
mvpt(
  LAV,
  path,
  data,
  showplots = FALSE,
  na.action = na.omit,
  subset = NULL,
  varcov = NULL,
  n = NULL
)
}
\arguments{
\item{LAV}{A pre-fitted SEM in lavaan syntax}

\item{path}{The SEM path to be tested, also in lavaan syntax}

\item{data}{A data frame to fit the given SEM}

\item{showplots}{Whether to show a figure containing all auto-generated SEMs}

\item{na.action}{SEM option}

\item{subset}{SEM option}

\item{varcov}{SEM option}

\item{n}{SEM option}
}
\value{
An object of class \code{mvpt} containing...
}
\description{
Given a model and path in lavaan syntax, plus data frame, this function auto-generates other models with the same path, fits each one, and tests for path values differences.
}
\examples{
# example code
\dontrun{
library(lavaan)
dat <- lavaan::HolzingerSwineford1939
LAV <- "visual  =~ x1 + x2 + x3
        textual =~ x4 + x5 + x6
        speed   =~ x7 + x8 + x9
        visual  ~ textual"
path <- "visual ~ textual"
out <- mvpt(LAV = LAV, path = path, data = dat)
out}
}
