% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{mvpt}
\alias{mvpt}
\title{Run MVP Test}
\usage{
mvpt(
  lavaan_model,
  path,
  data,
  showplots = FALSE,
  na.action = na.omit,
  subset = NULL,
  varcov = NULL,
  n = NULL
)
}
\arguments{
\item{lavaan_model}{A pre-fitted SEM in lavaan syntax.}

\item{path}{The path to be tested within the given SEM. This must also be in lavaan syntax (eg: Y~X).}

\item{data}{A data frame to fit the given SEM, and all others that may be auto-generated.}

\item{showplots}{Whether to show a parameter-free figure containing all the compared SEMs, with the user-supplied model always first. This figure is for quick inspection of model specification differences.}

\item{na.action}{lavaan option}

\item{subset}{lavaan option}

\item{varcov}{lavaan option}

\item{n}{lavaan option}
}
\value{
An object of class \code{mvpt} containing lavaan-fitted results, figures, and MVP test components.
}
\description{
Using a pre-fitted SEM and single path within that SEM, both in lavaan syntax, this function auto-generates other models with the same single path, though with differing relationships between other specified variables. This function then uses the supplied data to fit all models, followed by a test across models for significant value differences in the specified path.
}
\examples{
\dontrun{
data(PoliticalDemocracy, package = "lavaan")
lavaan_model <- 
  "
  ## regressions
  dem60 ~ ind60
  dem65 ~ ind60 + dem60
  ## latent variables
  ind60 =~ x1 + x2 + x3
  dem60 =~ y1 + a*y2 + b*y3 + c*y4
  dem65 =~ y5 + a*y6 + b*y7 + c*y8
  "
path <- "dem60 ~ ind60"
data <- PoliticalDemocracy
output <- mvpt(lavaan_model, path, data)
output}
}
