% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{mvpt}
\alias{mvpt}
\title{Run MVP Test}
\usage{
mvpt(lavaan_model, path, data, showplots = FALSE)
}
\arguments{
\item{lavaan_model}{A SEM in lavaan syntax.}

\item{path}{The path to be tested within the given SEM. This must also be in lavaan syntax (eg: Y~X).}

\item{data}{A data frame to fit the given SEM, and all others that may be auto-generated.}

\item{showplots}{Whether to show a parameter-free figure containing all the compared SEMs, with the user-supplied model always first. This figure is for quick inspection of model specification differences.}
}
\value{
An object of class \code{mvpt} containing lavaan-fitted results, figures, and MVP test components.
}
\description{
Using the model specification of a SEM and single path within that SEM, both in \pkg{lavaan} syntax, this function auto-generates multiple other models with the same single path by only using the graphical features of the given SEM. Auto-generated models will share the same fit statistics as the given SEM (Verma & Pearl, 1991), though suggest differing relationships between variables. After auto-generation, this function then uses the supplied data to fit all models using the same settings (limited to maximum likelihood estimation for now), followed by a chi-square test across models for significant value changes in the specified path. The given model will always be indexed first and appear as "M1" in the output.
}
\examples{
\dontrun{
data(PoliticalDemocracy, package = "lavaan")

lavaan_model <- 
  "
  ## latent variables
  ind60 =~ x1 + x2 + x3
  dem60 =~ y1 + y2 + y3 + y4
  dem65 =~ y5 + y6 + y7 + y8
  ## regressions
  dem60 ~ ind60
  dem65 ~ ind60 + dem60
  "
path <- "dem60 ~ ind60"
mvpt <- mvpt(lavaan_model, path, data = PoliticalDemocracy, showplots = TRUE)
mvpt}
}
