% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{mvptRank}
\alias{mvptRank}
\title{Rank Models by Path Value}
\usage{
mvptRank(output, top)
}
\arguments{
\item{output}{Output from using mvpt function}

\item{top}{Number of top models to include}
}
\value{
An object of class...
}
\description{
Follow-up function to rank and view models with largest path values.
}
\examples{
# example code
\dontrun{
out <- mvpt(LAV = LAV, path = path, data = dat)
mvptRank(out, top = 3)}
}
