% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{mvptRank}
\alias{mvptRank}
\title{Rank Models by Path Value}
\usage{
mvptRank(mvpt_output, top)
}
\arguments{
\item{mvpt_output}{Output from using mvpt() function.}

\item{top}{Number of top models to include in report.}
}
\value{
An object of class...
}
\description{
Follow-up function to rank and view models with largest path values (updated rank by absolute value pending).
}
\examples{
\dontrun{
mvpt_output <- mvpt(lavaan_model, path, data)
mvptRank(mvpt_output, top = 3)}
}
