% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvptRank.R
\name{mvptRank}
\alias{mvptRank}
\title{mvpt() Follow-up: Rank Models by Path Value}
\usage{
mvptRank(mvpt_output)
}
\arguments{
\item{mvpt_output}{Output from using mvpt() function.}
}
\value{
An object of class...
}
\description{
Follow-up function to rank and view all models by path value.
}
\examples{
\dontrun{
mvpt_output <- mvpt(lavaan_model, path, data)
mvptRank(mvpt_output)}
}
