% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvptZoom.R
\name{mvptZoom}
\alias{mvptZoom}
\title{mvpt() Follow-up: View Single Model}
\usage{
mvptZoom(mvpt_output, index)
}
\arguments{
\item{mvpt_output}{Output from mvpt().}

\item{index}{Model index for list of compared model, both given and auto-generated. Model 1 is always the given model.}
}
\description{
Follow-up function to see figure and lavaan output for a single SEM.
}
\examples{
\dontrun{
mvpt_output <- mvpt(lavaan_model, path, data)
mvptZoom(mvpt_output, index = 5)}
}
