% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{mvptZoom}
\alias{mvptZoom}
\title{See MVP Test Results for a Single Model}
\usage{
mvptZoom(output, M)
}
\arguments{
\item{output}{Output from using mvpt function}

\item{M}{Model index}
}
\value{
An object of class...
}
\description{
Follow-up function to see figure and lavaan results for a user-specified model.
}
\examples{
# example code
\dontrun{
out <- mvpt(LAV = LAV, path = path, data = dat)
mvptZoom(out, M = 5)}
}
