% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpt.R
\name{mvptZoom}
\alias{mvptZoom}
\title{View Single Model}
\usage{
mvptZoom(mvpt_output, index)
}
\arguments{
\item{mvpt_output}{Output from using mvpt() function.}

\item{index}{Model index for list of compared model, both given and auto-generated. Model 1 is always the given model.}
}
\value{
An object of class...
}
\description{
Follow-up function to see figure and lavaan results for a single model.
}
\examples{
\dontrun{
mvpt_output <- mvpt(lavaan_model, path, data)
mvptZoom(mvpt_output, index = 5)}
}
